+++ location-tracker/include/Storage.h
@@ -1,121 +1,25 @@
+#ifndef __STORAGE_H__
+#define __STORAGE_H__
+
 #include <FS.h>
+#include <SD_MMC.h>
 
-void listDir(fs::FS &fs, const char *dirname, uint8_t levels)
-{
-    Serial.printf("Listing directory: %s\n", dirname);
+void listDir(fs::FS &fs, const char *dirname, uint8_t levels);
 
-    File root = fs.open(dirname);
-    if (!root) {
-        Serial.println("Failed to open directory");
-        return;
-    }
-    if (!root.isDirectory()) {
-        Serial.println("Not a directory");
-        return;
-    }
+void createDir(fs::FS &fs, const char *path);
 
-    File file = root.openNextFile();
-    while (file) {
-        if (file.isDirectory()) {
-            Serial.print("  DIR : ");
-            Serial.println(file.name());
-            if (levels) {
-                listDir(fs, file.path(), levels - 1);
-            }
-        } else {
-            Serial.print("  FILE: ");
-            Serial.print(file.name());
-            Serial.print("  SIZE: ");
-            Serial.println(file.size());
-        }
-        file = root.openNextFile();
-    }
-}
+void removeDir(fs::FS &fs, const char *path);
 
-void createDir(fs::FS &fs, const char *path)
-{
-    Serial.printf("Creating Dir: %s\n", path);
-    if (fs.mkdir(path)) {
-        Serial.println("Dir created");
-    } else {
-        Serial.println("mkdir failed");
-    }
-}
+void readFile(fs::FS &fs, const char *path);
 
-void removeDir(fs::FS &fs, const char *path)
-{
-    Serial.printf("Removing Dir: %s\n", path);
-    if (fs.rmdir(path)) {
-        Serial.println("Dir removed");
-    } else {
-        Serial.println("rmdir failed");
-    }
-}
+void writeFile(fs::FS &fs, const char *path, const char *message);
 
-void readFile(fs::FS &fs, const char *path)
-{
-    Serial.printf("Reading file: %s\n", path);
+void appendFile(fs::FS &fs, const char *path, const char *message);
 
-    File file = fs.open(path);
-    if (!file) {
-        Serial.println("Failed to open file for reading");
-        return;
-    }
+void renameFile(fs::FS &fs, const char *path1, const char *path2);
 
-    Serial.print("Read from file: ");
-    while (file.available()) {
-        Serial.write(file.read());
-    }
-}
+void deleteFile(fs::FS &fs, const char *path);
 
-void writeFile(fs::FS &fs, const char *path, const char *message)
-{
-    Serial.printf("Writing file: %s\n", path);
+void init_sd();
 
-    File file = fs.open(path, FILE_WRITE);
-    if (!file) {
-        Serial.println("Failed to open file for writing");
-        return;
-    }
-    if (file.print(message)) {
-        Serial.println("File written");
-    } else {
-        Serial.println("Write failed");
-    }
-}
-
-void appendFile(fs::FS &fs, const char *path, const char *message)
-{
-    Serial.printf("Appending to file: %s\n", path);
-
-    File file = fs.open(path, FILE_APPEND);
-    if (!file) {
-        Serial.println("Failed to open file for appending");
-        return;
-    }
-    if (file.print(message)) {
-        Serial.println("Message appended");
-    } else {
-        Serial.println("Append failed");
-    }
-}
-
-void renameFile(fs::FS &fs, const char *path1, const char *path2)
-{
-    Serial.printf("Renaming file %s to %s\n", path1, path2);
-    if (fs.rename(path1, path2)) {
-        Serial.println("File renamed");
-    } else {
-        Serial.println("Rename failed");
-    }
-}
-
-void deleteFile(fs::FS &fs, const char *path)
-{
-    Serial.printf("Deleting file: %s\n", path);
-    if (fs.remove(path)) {
-        Serial.println("File deleted");
-    } else {
-        Serial.println("Delete failed");
-    }
-}
+#endif // __STORAGE_H__
