+++ location-tracker/include/power.h
@@ -1,0 +1,7 @@
+#include <Arduino.h>
+
+#include "utilities.h"
+#define XPOWERS_CHIP_AXP2101
+#include "XPowersLib.h"
+
+void init_pmu();
